@@ -7,9 +7,9 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
 # followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
+# Both the version and the build number may be overridden in flutter
 # build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
+# In Android, build-name is used as versionName while build-number is used as versionCode.
 # Read more about Android versioning at https://developer.android.com/studio/publish/versioning
 # In iOS, build-name is used as CFBundleShortVersionString while build-number is used as CFBundleVersion.
 # Read more about iOS versioning at
@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.7.0
+  sdk: ">=3.7.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,16 +30,25 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  geolocator: ^13.0.3
+  geocoding: ^3.0.0  # (Optional) For getting address from coordinates
   google_maps_flutter: ^2.5.0  # Ensure you use the latest version
+  google_places_flutter: ^2.0.5
+  font_awesome_flutter: ^10.6.0
+  uuid: ^4.5.1
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
+  flutter_polyline_points: ^2.1.0
+  flutter_google_maps_webservices: ^1.1.1
+  flutter_dotenv: ^5.2.1
+  google_fonts: ^6.2.1
+  async: ^2.12.0
+  geolocator_android: ^4.6.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  geolocator: ^13.0.3
-  geocoding: ^3.0.0  # (Optional) For getting address from coordinates
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
@@ -61,7 +70,8 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-       - assets/
+      - lib/assets/.env 
+      - assets/
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
 
